requirements.txt
python-kasa
structlog
asyncio
python-dotenv
uvicorn
fastapi
prometheus-fastapi-instrumentator
pydash
pydotted